requirements.txt
pandas
openai
numpy
python-dotenv
selenium